+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
  - Display proper error messages in certain cases when corrupted VDF files are found
  - Fix crash caused by appmanifest files that can't be read due to insufficient permissions
  - Fix crash caused by non-Proton compatibility tool being enabled for the selected app
+ - Fix erroneous warning when Steam library is inside a case-insensitive file system
 
 ## [1.5.2] - 2021-06-09
 ### Fixed